Add dmax keyword in fix_badpix()
Solution proposed following ticket #9. The bug in this ticket cannot
be reproduced, however it seems that the problem does not occur in the
current implementation. The dmax keyword provides more control. The
default value is set to 10 pixel, which is much more realistic than
the previous value of 100!

Ticket #9 is now closed.